worker/requirements.txt
openai-whisper
torch
torchaudio
requests
pyyaml
tqdm
